serving/requirements.txt
streamlit
snowflake-connector-python
cryptography
plotly
